aplicando princípio ocp
