+++ [OCP]-app_etl/src/extrator/Arquivo.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace src\extrator;
+
+class Arquivo {
+
+    private $dados = array();
+
+    public function setDados(string $nome, string $cpf, string $email): void {
+        array_push(
+            $this->dados,
+            [
+                'nome' => iconv('iso-8859-1', 'utf-8', $nome), //correção de caracter especial
+                'cpf' => $cpf, 
+                'email' => $email
+            ]);
+    }
+
+    public function getDados(): array {
+        return $this->dados;
+    }
+  
+}
