+++ [OCP]-app_etl/src/extrator/Txt.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace src\extrator;
+
+class Txt extends Arquivo{
+
+    public function lerArquivo(string $caminho): array {
+        
+        $handle = fopen($caminho, 'r');
+
+        while(!feof($handle)) {
+            $linha = fgets($handle);  // o ponteiro de leitura do arquivo é incrementado
+
+            $this->setDados(
+                substr($linha, 11, 30),  // nome
+                substr($linha, 0, 11),  // cpf
+                substr($linha, 41, 50)  // email
+            );
+        }
+
+        fclose($handle);
+
+        return $this->getDados();
+    }
+
+}
